+++ frontend/app/types/session/session.ts
@@ -115,12 +115,8 @@ export default Record(
         .filter(({ type, time }) => type !== TYPES.CONSOLE && time <= durationSeconds);
 
       let resources = List(session.resources).map(Resource);
-      // this code shoud die.
-      const firstResourceTime = resources
-        .map((r) => r.time)
-        .reduce((a, b) => Math.min(a, b), Number.MAX_SAFE_INTEGER);
       resources = resources
-        .map((r) => r.set('time', r.time - firstResourceTime))
+        .map((r) => r.set('time', Math.max(0, r.time - startedAt)))
         .sort((r1, r2) => r1.time - r2.time);
       const missedResources = resources.filter(({ success }) => !success);
 
@@ -171,7 +167,6 @@ export default Record(
         ),
         userDisplayName:
           session.userId || session.userAnonymousId || session.userID || 'Anonymous User',
-        firstResourceTime,
         issues: issuesList,
         sessionId: sessionId || sessionID,
         userId: session.userId || session.userID,