+++ tracker/tracker/src/webworker/MessageEncoder.gen.ts
@@ -86,6 +86,10 @@ export default class MessageEncoder extends PrimitiveEncoder {
       return  this.uint(msg[1]) && this.uint(msg[2]) 
     break
     
+    case Messages.Type.NetworkRequest:
+      return  this.string(msg[1]) && this.string(msg[2]) && this.string(msg[3]) && this.string(msg[4]) && this.string(msg[5]) && this.uint(msg[6]) && this.uint(msg[7]) && this.uint(msg[8]) 
+    break
+    
     case Messages.Type.ConsoleLog:
       return  this.string(msg[1]) && this.string(msg[2]) 
     break